Adjust package structure
