+++ spring-cloud-stream-schema/src/main/java/org/springframework/cloud/stream/schema/SchemaRegistrationResponse.java
@@ -14,30 +14,30 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.schema.avro;
-
-import org.apache.avro.Schema;
+package org.springframework.cloud.stream.schema;
 
 /**
- * @author Vinicius Carvalho
  * @author Marius Bogoevici
  */
-public interface SchemaRegistryClient {
+public class SchemaRegistrationResponse {
 
-	/**
-	 * Registers a schema with the remote repository returning the unique identifier associated with this schema.
-	 * version
-	 * @param subject the full name of the schena
-	 * @param schema
-	 * @return a {@link SchemaRegistrationResponse} representing the result of the operation
-	 */
-	SchemaRegistrationResponse register(String subject, Schema schema);
+	private long id;
 
-	/**
-	 * Retrieves a schema by its identifier.
-	 * @param id
-	 * @return
-	 */
-	Schema fetch(SchemaReference id);
+	private SchemaReference schemaReference;
 
+	public long getId() {
+		return this.id;
+	}
+
+	public void setId(long id) {
+		this.id = id;
+	}
+
+	public SchemaReference getSchemaReference() {
+		return this.schemaReference;
+	}
+
+	public void setSchemaReference(SchemaReference schemaReference) {
+		this.schemaReference = schemaReference;
+	}
 }
